+++ lib/amap_map.dart
@@ -30,5 +30,4 @@ export 'package:amap_map/src/types/types.dart';
 part 'src/amap_initializer.dart';
 part 'src/amap_controller.dart';
 part 'src/amap_widget.dart';
-part 'src/amap_loader.dart';
 part 'src/utils/location_utils.dart';