+++ lib/src/extension/extensions.dart
@@ -1,1 +1,0 @@
-export 'buildin/info_window_extension.dart';